# build: fix pd rebuild on testnet dir