Fixes build caching for pd. Follow up to #2292. Ensures that the testnet directory referenced by the pd build script is a fullpath, rather than a dirname. The dirname didn't resolve to a valid relpath from the perspective of the build, so cargo was always rebuilding pd, even when no code had changed.

For reference, the command I used to identify the always-marked-dirty dir build dep was:

  CARGO_LOG=cargo::core::compiler::fingerprint=info cargo build --release --bin pd

The speedup on hot builds of pd is ~100x, 30s -> 0.3s on my machine. Benchmarks from local runs:

On main:

```
❯ hyperfine --warmup 2 --runs 5 'cargo build --release --bin pd' 
Benchmark 1: cargo build --release --bin pd
  Time (mean ± σ):     32.689 s ±  0.244 s    [User: 143.447 s, System: 3.406 s]
  Range (min … max):   32.394 s … 33.009 s    5 runs

```

On this feature branch:

```
❯ hyperfine --warmup 2 --runs 5 'cargo build --release --bin pd' 
Benchmark 1: cargo build --release --bin pd
  Time (mean ± σ):     305.5 ms ±   2.2 ms    [User: 236.6 ms, System: 67.0 ms]
  Range (min … max):   302.5 ms … 307.3 ms    5 runs
```