@conorsch 2023-05-01
I tacked on one more commit so that a general `cargo build --release` rebuild is speedy now, too.